requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
requests==2.32.4
faiss-cpu==1.12.0
numpy==2.3.2
pydantic==2.11.7
scikit-learn==1.7.1
pypdf==6.0.0
python-docx==1.2.0
python-multipart==0.0.20
pytest==8.4.1
